[MODIFY]: 기존 검색어 레디스 서비스를 레디스 서비스와 일반 서비스 크래스로 분리
- 레디스 서비스 클래스에서 사용자를 조회하는 로직을 분리
- 컨트롤러에서 다이렉트로 레디스 서비스를 사용하기보다 일반 서비스를 거쳐 사용하도록 변경
