+++ src/main/java/com/carrot/carrotmarketclonecoding/board/service/SearchKeywordService.java
@@ -4,15 +4,8 @@ import java.util.List;
 import java.util.Set;
 
 public interface SearchKeywordService {
-    void addSearchKeywordRank(String keyword);
-
     Set<String> getTopSearchKeywords();
-
-    void addRecentSearchKeywords(Long memberId, String keyword);
-
-    List<String> getRecentSearchKeywords(Long memberId);
-
-    void removeRecentSearchKeyword(Long memberId, String keyword);
-
-    void removeAllRecentSearchKeywords(Long memberId);
+    List<String> getRecentSearchKeywords(Long authId);
+    void removeRecentSearchKeyword(Long authId, String keyword);
+    void removeAllRecentSearchKeywords(Long authId);
 }
